% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_clusters_hmap.R
\name{all_clusters_hmap}
\alias{all_clusters_hmap}
\title{Create a Heatmap for All Clusters}
\usage{
all_clusters_hmap(
  full_clusterdf,
  value_type = "Padj",
  x_labels = NULL,
  title = NULL
)
}
\arguments{
\item{full_clusterdf}{A dataframe containing the full cluster data.}

\item{value_type}{A string specifying the type of value to use for the heatmap (default is "Padj").}

\item{x_labels}{A vector of strings specifying the labels for the x-axis (default is NULL, which uses the original column names).}

\item{title}{A string specifying the title of the heatmap (default is NULL, which generates a default title).}
}
\value{
An interactive heatmaply heatmap.
}
\description{
This function generates a heatmap for all clusters based on the specified value type (e.g., Padj or Pvalue).
}
