% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_correlation.R
\name{cluster_correlation_hmap}
\alias{cluster_correlation_hmap}
\title{Create a Correlation Heatmap for a Specific Cluster}
\usage{
cluster_correlation_hmap(
  final_clusters,
  distance_matrix,
  cluster_number,
  merged_df
)
}
\arguments{
\item{final_clusters}{A dataframe containing the final cluster data.}

\item{distance_matrix}{A matrix representing the distances between terms.}

\item{cluster_number}{An integer cluster id, as found in \code{final_clusters$Cluster}.}

\item{merged_df}{A dataframe with all terms used to map term indices to names.}
}
\value{
An interactive heatmaply heatmap.
}
\description{
This function generates a correlation heatmap for a specific cluster based on the provided distance matrix.
}
\examples{
\donttest{
cluster_result <- readRDS(system.file("extdata", "cluster_result.rds",
                                      package = "richCluster"))
chmap <- cluster_correlation_hmap(cluster_result$final_clusters,
                                  cluster_result$distance_matrix,
                                  cluster_number = 1,
                                  cluster_result$merged_df)
chmap
}
}
