% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_correlation.R
\name{cluster_network}
\alias{cluster_network}
\title{Create a Network Graph for a Specific Cluster}
\usage{
cluster_network(final_clusters, distance_matrix, cluster_number, merged_df)
}
\arguments{
\item{final_clusters}{A dataframe containing the final cluster data.}

\item{distance_matrix}{A matrix representing the distances between terms.}

\item{cluster_number}{An integer cluster id, as found in \code{final_clusters$Cluster}.}

\item{merged_df}{A dataframe with all terms used to map term indices to names.}
}
\value{
An interactive networkD3 network graph.
}
\description{
This function generates a network graph for a specific cluster based on the provided distance matrix.
The opacity and length of the edges correspond to the given distance_metric (eg, kappa) score similarity
between terms, which is based on shared gene content.
}
\examples{
\donttest{
cluster_result <- readRDS(system.file("extdata", "cluster_result.rds",
                                      package = "richCluster"))
net <- cluster_network(cluster_result$final_clusters,
                       cluster_result$distance_matrix,
                       cluster_number = 1,
                       cluster_result$merged_df)
net
}
}
