% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_enrichment_results.R
\name{format_colnames}
\alias{format_colnames}
\title{Format Column Names for Merging}
\usage{
format_colnames(colnames)
}
\arguments{
\item{colnames}{A character vector of column names to be standardized.}
}
\value{
A character vector of standardized column names.
}
\description{
This function maps a vector of column names to standardized names
for "GeneID", "Pvalue", and "Padj" based on known variations.
}
