% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_merged_seeds.R
\name{make_full_clusterdf}
\alias{make_full_clusterdf}
\title{Make Full Cluster Dataframe}
\usage{
make_full_clusterdf(final_clusters, merged_richsets)
}
\arguments{
\item{final_clusters}{The final clusters data frame containing a 'ClusterIndices' column with
comma-delimited indices corresponding to terms in the merged_richsets df.}

\item{merged_richsets}{The merged_richsets data frame containing the original
combined Pvalue, Padj, and GeneID information (from merge_richsets function)}
}
\value{
A reconstructed full_cluster dataframe, adds a 'Cluster' column to the merged_richsets
}
\description{
Use the final clusters and merged richsets to create a full cluster data frame
containing all Pvalue, Padj, and GeneID information for each term in each cluster.
}
