% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_clusters_hmap.R
\name{make_summary_data}
\alias{make_summary_data}
\title{Preprocess the data for heatmap plotting}
\usage{
make_summary_data(full_clusterdf, value_type = "Padj")
}
\arguments{
\item{value_type}{The type of value to use for heatmap plotting (e.g., "Padj")}

\item{data}{The full cluster data frame containing clusterNumber and value columns}

\item{summary_metric}{The summary metric to use for summarizing the value columns (e.g., mean)}
}
\description{
Preprocess the data for heatmap plotting
}
