% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_richsets.R
\name{merge_richsets}
\alias{merge_richsets}
\title{Merge List of Enrichment Results}
\usage{
merge_richsets(richsets)
}
\arguments{
\item{richsets}{A list of richset dataframes containing columns c('Term', 'GeneID', 'Pvalue', 'Padj')}
}
\value{
A single merged richset dataframe with all original columns
        suffixed with the index of the richset, with new columns 'GeneID', 'Pvalue',
        'Padj' containing the merged values.
}
\description{
This function merges multiple enrichment results ('richsets') into a single richset by
combining unique GeneID elements across each richset, and averaging Pvalue / Padj
values for each term across all richsets.
}
