% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RichCluster}
\alias{RichCluster}
\title{RichCluster}
\usage{
RichCluster(
  distanceMetric,
  distanceCutoff,
  mergeStrategy,
  membershipCutoff,
  termNameColumn,
  geneIDColumn,
  PvalueColumn
)
}
\arguments{
\item{distanceMetric}{A string specifying the distance metric to use (e.g., "kappa").}

\item{distanceCutoff}{A double specifying the distance cutoff value.}

\item{mergeStrategy}{A string specifying the merge strategy to use (e.g., "DAVID").}

\item{membershipCutoff}{A double specifying the membership cutoff value (between 0 and 1).}

\item{termNameColumn}{A CharacterVector containing term names.}

\item{geneIDColumn}{A CharacterVector containing gene IDs.}

\item{PvalueColumn}{A NumericVector containing p-values.}
}
\value{
An R List object containing the following elements:
\itemize{
  \item{DistanceMatrix: The distance matrix used in clustering.}
  \item{SeedMap: The initial seed map of clusters.}
  \item{FilteredSeeds: The filtered seed map after applying the merge strategy.}
  \item{MergedSeeds: The final cluster list after merging seeds.}
}
}
\description{
This function clusters terms within an enrichment result with based on gene 
similarity using specified distance metrics and merging strategies. It 
integrates the following steps:
\enumerate{
  \item{Initializes a ClusterManager with input gene data.}
  \item{Calculates distance scores using the specified distance metric and cutoff.}
  \item{Filters seeds based on the provided merging strategy and membership cutoff.}
  \item{Merges seeds based on the merging strategy.}
}
}
\examples{
# Example usage
result <- RichCluster("kappa", 0.5, "DAVID", 0.7, termNames, geneIDs, pValues)
distanceMatrix <- result$DistanceMatrix
seedMap <- result$SeedMap
filteredSeeds <- result$FilteredSeeds
mergedSeeds <- result$MergedSeeds

}
