% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny_app.R
\name{run_shiny_app}
\alias{run_shiny_app}
\title{Launch the richCluster Shiny app}
\usage{
run_shiny_app(...)
}
\value{
No return value.
}
\description{
Starts the app shipped in inst/application.
}
\examples{
\dontshow{if (interactive() && requireNamespace("shiny", quietly = TRUE)) withAutoprint(\{ # examplesIf}
# run_shiny_app()
\dontshow{\}) # examplesIf}
}
