% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_clusters_hmap.R
\name{all_clusters_hmap}
\alias{all_clusters_hmap}
\title{Create a Heatmap for All Clusters}
\usage{
all_clusters_hmap(full_clusterdf, value_type = "Padj")
}
\arguments{
\item{full_clusterdf}{A dataframe containing the full cluster data.}

\item{value_type}{A string specifying the type of value to use for the heatmap (default is "Padj").}
}
\value{
An interactive heatmaply heatmap.
}
\description{
This function generates a heatmap for all clusters based on the specified value type (e.g., Padj or Pvalue).
}
