% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_correlation.R
\name{cluster_correlation_hmap}
\alias{cluster_correlation_hmap}
\title{Create a Correlation Heatmap for a Specific Cluster}
\usage{
cluster_correlation_hmap(final_clusters, distance_matrix, cluster_number, merged_df)
}
\arguments{
\item{final_clusters}{A dataframe containing the final cluster data.}

\item{distance_matrix}{A matrix representing the distances between terms.}

\item{cluster_number}{An integer specifying the cluster number to visualize.}

\item{merged_df}{A dataframe with all terms used to map term indices to names.}
}
\value{
An interactive heatmaply heatmap.
}
\description{
This function generates a correlation heatmap for a specific cluster based on the provided distance matrix.
}
