% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_correlation.R
\name{cluster_network}
\alias{cluster_network}
\title{Create a Network Graph for a Specific Cluster}
\usage{
cluster_network(final_clusters, distance_matrix, cluster_number)
}
\arguments{
\item{final_clusters}{A dataframe containing the final cluster data.}

\item{distance_matrix}{A matrix representing the distances between terms.}

\item{cluster_number}{An integer specifying the cluster number to visualize.}
}
\value{
An interactive networkD3 network graph.
}
\description{
This function generates a network graph for a specific cluster based on the provided distance matrix.
The opacity and length of the edges correspond to the given distance_metric (eg, kappa) score similarity 
between terms, which is based on shared gene content.
}
