% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_network_graphs_plotly.R
\name{compare_network_graphs_plotly}
\alias{compare_network_graphs_plotly}
\title{Compare Network Graphs using Plotly}
\usage{
compare_network_graphs_plotly(cluster_result, cluster_num, pval_names)
}
\arguments{
\item{cluster_result}{The result from the clustering function.}

\item{cluster_num}{The cluster number to plot.}

\item{pval_names}{A list of p-value names to compare.}
}
\value{
A plotly object.
}
\description{
This function creates a side-by-side comparison of network graphs for a single
cluster using different p-value types.
}
