% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_df.R
\name{export_df}
\alias{export_df}
\title{Export Cluster Result as Dataframe}
\usage{
export_df(cluster_result)
}
\arguments{
\item{cluster_result}{The cluster_result object from cluster()}
}
\value{
A data.frame view of the clustering
}
\description{
Returns a comprehensive dataframe containing all the different terms in all clusters.
}
