% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_network.R
\name{full_network}
\alias{full_network}
\title{Create a Network Graph for the Entire Distance Matrix}
\usage{
full_network(cluster_result)
}
\arguments{
\item{cluster_result}{Cluster result named list from RichCluster::cluster()}
}
\value{
An interactive networkD3 network graph.
}
\description{
This function generates a network graph for the entire distance matrix.
}
