% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_enrichment_results.R
\name{merge_enrichment_results}
\alias{merge_enrichment_results}
\title{Merge List of Enrichment Results}
\usage{
merge_enrichment_results(enrichment_results)
}
\arguments{
\item{enrichment_results}{A list of geneset dataframes containing columns c('Term', 'GeneID', 'Pvalue', 'Padj')}
}
\value{
A single merged geneset dataframe with all original columns
        suffixed with the index of the geneset, with new columns 'GeneID', 'Pvalue',
        'Padj' containing the merged values.
}
\description{
This function merges multiple enrichment results ('enrichment_results') into a single dataframe by
combining unique GeneID elements across each geneset, and averaging Pvalue / Padj
values for each term across all enrichment_results.
}
