% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network_graph.R
\name{plot_network_graph}
\alias{plot_network_graph}
\title{Plot Network Graph for a Cluster}
\usage{
plot_network_graph(
  cluster_result,
  cluster_num,
  distance_matrix,
  valuetype_list
)
}
\arguments{
\item{cluster_result}{The result from the clustering function.}

\item{cluster_num}{The cluster number to plot.}

\item{distance_matrix}{The distance matrix used for clustering.}

\item{valuetype_list}{A list of value types (e.g., "Pvalue_1", "Padj_1") to use for node coloring.}
}
\value{
A plot object.
}
\description{
This function visualizes a single cluster as a network graph.
}
