% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richCluster-package.R
\docType{package}
\name{richCluster-package}
\alias{richCluster}
\alias{richCluster-package}
\title{richCluster: clustering and visualization utilities}
\description{
Tools for clustering enriched terms, building correlation networks,
and producing interactive heatmaps and network views.
}
\author{
\strong{Maintainer}: Junguk Hur \email{hurlabshared@gmail.com} (\href{https://orcid.org/0000-0002-0736-2149}{ORCID})

Authors:
\itemize{
  \item Sarah Hong
  \item Jane Kim
}

}
\keyword{internal}
