% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richCluster-package.R
\docType{package}
\name{richCluster-package}
\alias{richCluster-package}
\title{richCluster: An R Package for Clustering Analysis}
\description{
This package provides tools to perform clustering... (add a short description).
}
\author{
\strong{Maintainer}: Junguk Hur \email{junguk.hur@med.und.edu}

Authors:
\itemize{
  \item Sarah Hong \email{ssh2198@columbia.edu}
  \item Jane Kim \email{j6kim2026@chadwickschool.org}
}

}
