% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{runRichCluster}
\alias{runRichCluster}
\title{Run clustering in C++ backend}
\usage{
runRichCluster(
  terms,
  geneIDs,
  distanceMetric,
  distanceCutoff,
  linkageMethod,
  linkageCutoff
)
}
\arguments{
\item{terms}{Character vector of term names}

\item{geneIDs}{Character vector of geneIDs}

\item{distanceMetric}{e.g. "kappa"}

\item{distanceCutoff}{numeric between 0 and 1}

\item{linkageMethod}{e.g. "average"}

\item{linkageCutoff}{numeric between 0 and 1}
}
\description{
Run clustering in C++ backend
}
